back/requirements.txt
Package            Version
------------------ -----------
alembic            1.13.2
annotated-types    0.7.0
beautifulsoup4     4.12.3
blinker            1.8.2
blis               0.7.11
catalogue          2.0.10
certifi            2024.8.30
cffi               1.17.1
charset-normalizer 3.3.2
ci-info            0.3.0
click              8.1.7
cloudpathlib       0.19.0
confection         0.1.5
configobj          5.0.8
configparser       7.1.0
cryptography       43.0.1
cymem              2.0.8
etelemetry         0.3.1
filelock           3.16.1
fitz               0.0.1.dev2
Flask              3.0.3
Flask-Cors         5.0.0
Flask-Dance        7.1.0
Flask-Login        0.6.3
Flask-Migrate      4.0.7
Flask-SQLAlchemy   3.1.1
fr-core-news-md    3.7.0
fr-core-news-sm    3.7.0
greenlet           3.1.0
httplib2           0.22.0
idna               3.9
isodate            0.6.1
itsdangerous       2.2.0
Jinja2             3.1.4
langcodes          3.4.0
language_data      1.2.0
looseversion       1.3.0
lxml               5.3.0
Mako               1.3.5
marisa-trie        1.2.0
markdown-it-py     3.0.0
MarkupSafe         2.1.5
mdurl              0.1.2
murmurhash         1.0.10
networkx           3.3
nibabel            5.2.1
nipype             1.8.6
numpy              1.26.4
oauthlib           3.2.2
packaging          24.1
pandas             2.2.2
pathlib            1.0.1
pdfminer.six       20231228
pdfplumber         0.11.4
pillow             10.4.0
pip                24.2
preshed            3.0.9
prov               2.0.1
psycopg2-binary    2.9.9
pycparser          2.22
pydantic           2.9.1
pydantic_core      2.23.3
pydot              3.0.1
Pygments           2.18.0
PyMuPDF            1.24.10
PyMuPDFb           1.24.10
pyparsing          3.1.4
pypdfium2          4.30.0
python-dateutil    2.9.0.post0
python-dotenv      1.0.1
pytz               2024.2
pyxnat             1.6.2
rdflib             6.3.2
requests           2.32.3
requests-oauthlib  2.0.0
rich               13.8.1
scipy              1.14.1
setuptools         68.2.0
shellingham        1.5.4
simplejson         3.19.3
six                1.16.0
smart-open         7.0.4
soupsieve          2.6
spacy              3.7.6
spacy-legacy       3.0.12
spacy-loggers      1.0.5
SQLAlchemy         2.0.34
srsly              2.4.8
thinc              8.2.5
tqdm               4.66.5
traits             6.3.2
typer              0.12.5
typing_extensions  4.12.2
tzdata             2024.1
urllib3            2.2.3
URLObject          2.4.3
wasabi             1.1.3
weasel             0.4.1
Werkzeug           3.0.4
wheel              0.41.2
wrapt              1.16.0
